[FIXED MIFOS-2410] Patch from Jakub Sławiński. https://mifos.dev.java.net/nonav/issues/showattachment.cgi/1209/issue_2410_v3.1.patch https://mifos.dev.java.net/issues/show_bug.cgi?id=2410 http://thread.gmane.org/gmane.comp.finance.mifos.devel/6898
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@16209 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ application/src/main/java/org/mifos/application/customer/business/CustomerCustomFieldEntity.java
@@ -34,13 +34,13 @@ public class CustomerCustomFieldEntity extends PersistentObject {
     /*
      * Reference to a {@link CustomFieldDefinitionEntity}
      */
-    private final Integer fieldId;
+    private final Short fieldId;
 
     private String fieldValue;
 
     private final CustomerBO customer;
 
-    public CustomerCustomFieldEntity(Integer fieldId, String fieldValue, CustomerBO customer) {
+    public CustomerCustomFieldEntity(Short fieldId, String fieldValue, CustomerBO customer) {
         this.fieldId = fieldId;
         this.fieldValue = fieldValue;
         this.customer = customer;
@@ -58,7 +58,7 @@ public class CustomerCustomFieldEntity extends PersistentObject {
         this.customer = null;
     }
 
-    public Integer getFieldId() {
+    public Short getFieldId() {
         return fieldId;
     }
